refactor class structure
